SDIT-3054: ⬆️  Upgrade to spring boot 3.5.6 & hmpps plugin 9.0.2 (#344)
* SDIT-3054: ⬆️  Upgrade to spring boot 3.5.6 & hmpps plugin 9.0.2

* SDIT-3054: 👽️ Workaround for spring data jpa issue by returning java.lang.Integer instead

+++ src/test/java/uk/gov/justice/digital/hmpps/cmd/api/service/NotificationServiceTest.kt
@@ -368,7 +368,7 @@ internal class NotificationServiceTest {
 
     @Test
     fun `Tidy notifications`() {
-      every { shiftNotificationRepository.deleteAllByShiftModifiedBefore(any()) } returns 4
+      every { shiftNotificationRepository.deleteAllByShiftModifiedBefore(any()) } returns 4 as Integer
 
       service.tidyNotification()
